# `xml_name_escape`: Optimize the common case by 2-3x