--- a/activesupport/test/core_ext/string_ext_test.rb
+++ b/activesupport/test/core_ext/string_ext_test.rb
@@ -1054,6 +1054,7 @@ class OutputSafetyTest < ActiveSupport::TestCase
     unsafe_char = ">"
     safe_char = "Á"
     safe_char_after_start = "3"
+    starting_with_dash = "-foo"
 
     assert_equal "_", ERB::Util.xml_name_escape(unsafe_char)
     assert_equal "_#{safe_char}", ERB::Util.xml_name_escape(unsafe_char + safe_char)
@@ -1074,6 +1075,8 @@ class OutputSafetyTest < ActiveSupport::TestCase
     common_dangerous_chars = "&<>\"' %*+,/;=^|"
     assert_equal "_" * common_dangerous_chars.size,
                  ERB::Util.xml_name_escape(common_dangerous_chars)
+
+    assert_equal "_foo", ERB::Util.xml_name_escape(starting_with_dash)
   end
 end